fixed: `transitionend` event no longer fires twice
